// src/app/shared/interfaces/business-type.interface.ts
export enum BusinessTypeEnum{
    CIVIL_INFRASTRUCTURE = 'Civil Infrastructure',
    CONSTRUCTION = 'Construction',
    ENERGY = 'Energy',
    ENGINEERING_ARCHITECTURE = 'Engineering/Architecture',
    GOVERNMENT = 'Government',
    MANUFACTURING = 'Manufacturing',
    OIL_GAS = 'Oil & Gas',
    PHARMACEUTICAL = 'Pharmaceutical',
    SURVEY_GEOSPATIAL = 'Survey & Geospatial',
    TELECOMMUNICATIONS = 'Telecommunications',
    OTHER = 'Other'
}